+++ apps/server/src/infrastructure/persistence/typeorm/entity/like/typeorm-like.entity.spec.ts
@@ -1,20 +1,20 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { DataSource, Repository } from "typeorm";
 import { TypeOrmModule } from "@nestjs/typeorm";
-import { typeormSqliteOptions } from "../config/typeorm-config";
 import { join } from "path";
-import { TypeormComment } from "./typeorm-comment.entity";
+import { TypeormLike } from "../like/typeorm-like.entity";
 import { DummyDatabaseHandler } from "@test-utils/persistence/dummy-database-handler";
+import { typeormSqliteOptions } from "../../config/typeorm-config";
 
 const parameters = {
-  testDbPath: join("db", "TypeormComment.sqlite"),
+  testDbPath: join("db", "TypeormLike.sqlite"),
   dummyDbPath: join("db", "dummy.sqlite"),
 };
 
-describe("TypeormComment", () => {
+describe("TypeormLike", () => {
   let module: TestingModule;
   let dataSource: DataSource;
-  let repository: Repository<TypeormComment>;
+  let repository: Repository<TypeormLike>;
   let testDatabaseHandler: DummyDatabaseHandler;
 
   beforeAll(async () => {
@@ -30,7 +30,7 @@ describe("TypeormComment", () => {
     }).compile();
 
     dataSource = module.get<DataSource>(DataSource);
-    repository = dataSource.getRepository(TypeormComment);
+    repository = dataSource.getRepository(TypeormLike);
 
     testDatabaseHandler = new DummyDatabaseHandler(dataSource);
 
@@ -43,16 +43,16 @@ describe("TypeormComment", () => {
 
   describe("save", () => {
     it("should save normally", async () => {
-      const comment = testDatabaseHandler.makeDummyComment();
-      const savedComment = await repository.save(comment);
-      expectEqualComment(comment, savedComment);
+      const like = await testDatabaseHandler.makeDummyLike();
+      const savedLike = await repository.save(like);
+      await expectEqualLike(like, savedLike);
     });
   });
 
   describe("delete", () => {
-    let target!: TypeormComment;
+    let target!: TypeormLike;
     beforeEach(async () => {
-      target = testDatabaseHandler.makeDummyComment();
+      target = await testDatabaseHandler.makeDummyLike();
       target = await repository.save(target);
     });
 
@@ -62,21 +62,20 @@ describe("TypeormComment", () => {
 
     it("should delete normally", async () => {
       await repository.delete(target.id);
-      const deletedComment = await repository.findOneBy({ id: target.id });
-      expect(deletedComment).toBeNull();
+
+      const deletedLike = await repository.findOneBy({ id: target.id });
+
+      expect(deletedLike).toBeNull();
     });
   });
 });
 
-function expectEqualComment(
-  comment: TypeormComment,
-  savedComment: TypeormComment,
-) {
-  expect(comment.id).toBe(savedComment.id);
-  expect(comment.type).toBe(savedComment.type);
-  expect(comment.text).toBe(savedComment.text);
-  expect(comment.contentId).toBe(savedComment.contentId);
-  expect(comment.createdDateTime).toBe(savedComment.createdDateTime);
-  expect(comment.updatedDateTime).toBe(savedComment.updatedDateTime);
-  expect(comment.deletedDateTime).toBe(savedComment.deletedDateTime);
+async function expectEqualLike(
+  like: TypeormLike,
+  savedLike: TypeormLike,
+): Promise<void> {
+  expect(like.id).toBe(savedLike.id);
+  expect((await like.content).id).toBe((await savedLike.content).id);
+  expect((await like.user).id).toBe((await savedLike.user).id);
+  expect(like.createdDateTime).toBe(savedLike.createdDateTime);
 }
